Remove pest
