+++ Cargo.toml
@@ -15,8 +15,6 @@ exclude = [
 [dependencies]
 nom = "8.0.0"
 nom_locate = "5.0.0"
-pest = "2.8"
-pest_derive = "2.8"
 
 [dev-dependencies]
 itertools = "0.14.0"
