+++ src/lib.rs
@@ -1,14 +1,14 @@
-#![doc = include_str!("../README.md")]
+use nom::branch::alt;
+use nom::bytes::complete::tag;
+use nom::character::complete::{alphanumeric1, newline, not_line_ending, space0, space1};
+use nom::combinator::{all_consuming, opt, recognize};
+use nom::multi::{many0, many1, separated_list1};
+use nom::sequence::{delimited, preceded, terminated};
+use nom::{IResult, Parser};
+use nom_locate::{LocatedSpan, position};
+use std::borrow::Borrow;
 
-pub mod nom;
-
-use pest::Parser;
-use pest::iterators::Pair;
-use pest_derive::Parser;
-
-#[derive(Parser)]
-#[grammar = "grammar.pest"]
-struct GrammarParser;
+type Span<'a> = LocatedSpan<&'a str>;
 
 #[derive(Debug, Eq, PartialEq, Clone)]
 pub struct Import {
@@ -18,129 +18,98 @@ pub struct Import {
 }
 
 impl Import {
-    pub fn new(imported_object: &str, line_number: u32, typechecking_only: bool) -> Self {
+    pub fn new<B: Borrow<str>>(
+        imported_object: B,
+        line_number: u32,
+        typechecking_only: bool,
+    ) -> Self {
         Self {
-            imported_object: imported_object.to_owned(),
+            imported_object: imported_object.borrow().to_owned(),
             line_number,
             typechecking_only,
         }
     }
 }
 
-#[derive(Debug)]
-struct ParseContext {
-    typechecking_only: bool,
+// TODO Remove grammar.pest
+
+pub fn parse_imports(s: &str) -> Result<Vec<Import>, String> {
+    let s = Span::new(s);
+    let (_, result) = all_consuming(many0(parse_import_statement_list))
+        .parse(s)
+        .map_err(|e| e.to_string())?;
+    Ok(result.into_iter().flatten().collect())
 }
 
-pub fn parse_imports(code: &str) -> Result<Vec<Import>, String> {
-    let pair = GrammarParser::parse(Rule::CODE, code)
-        .map_err(|e| format!("failed to parse: {}", e))?
-        .next()
-        .unwrap();
-
-    let mut context = ParseContext {
-        typechecking_only: false,
-    };
-
-    Ok(parse_pair(pair, &mut context))
+fn parse_import_statement_list(s: Span) -> IResult<Span, Vec<Import>> {
+    let (rest, result) = terminated(
+        separated_list1(
+            delimited(space0, tag(";"), space0),
+            alt((parse_import_statement, parse_from_import_statement)),
+        ),
+        (
+            opt(space0),
+            opt(tag(";")),
+            opt(space0),
+            opt(parse_comment),
+            opt(newline),
+        ),
+    )
+    .parse(s)?;
+    Ok((rest, result.into_iter().flatten().collect()))
 }
 
-fn parse_pair(pair: Pair<Rule>, context: &mut ParseContext) -> Vec<Import> {
-    match pair.as_rule() {
-        Rule::CODE
-        | Rule::FRAGMENT
-        | Rule::IMPORT_STATEMENT_LIST
-        | Rule::IF_TYPECHECKING_FRAGMENT => parse_inner_pairs(pair, context),
-        Rule::IF_TYPECHECKING | Rule::SINGLELINE_IF_TYPECHECKING => {
-            context.typechecking_only = true;
-            let imports = parse_inner_pairs(pair, context);
-            context.typechecking_only = false;
-            imports
-        }
-        Rule::SIMPLE_IMPORT_STATEMENT => parse_simple_import_statement(pair, context),
-        Rule::FROM_IMPORT_STATEMENT | Rule::MULTILINE_FROM_IMPORT_STATEMENT => {
-            parse_from_import_statement(pair, context)
-        }
-        Rule::WILDCARD_FROM_IMPORT_STATEMENT => parse_wildcard_from_import_statement(pair, context),
-        Rule::MULTILINE_STRING => {
-            vec![]
-        }
-        Rule::EOI => {
-            vec![]
-        }
-        _ => unreachable!("{:?}", pair.as_rule()),
-    }
+fn parse_import_statement(s: Span) -> IResult<Span, Vec<Import>> {
+    let (rest, result) = preceded(
+        (tag("import"), space1),
+        separated_list1(
+            delimited(space0, tag(","), space0),
+            terminated(
+                (position, parse_module),
+                opt((space1, tag("as"), space1, parse_identifier)),
+            ),
+        ),
+    )
+    .parse(s)?;
+    Ok((
+        rest,
+        result
+            .into_iter()
+            .map(|(span, module_name)| Import::new(module_name, span.location_line(), false))
+            .collect(),
+    ))
 }
 
-fn parse_inner_pairs(pair: Pair<Rule>, context: &mut ParseContext) -> Vec<Import> {
-    pair.into_inner().fold(vec![], |mut imports, inner_pair| {
-        imports.extend(parse_pair(inner_pair, context));
-        imports
-    })
+fn parse_from_import_statement(s: Span) -> IResult<Span, Vec<Import>> {
+    let (rest, result) = (
+        (tag("from"), space1),
+        parse_module,
+        (space1, tag("import"), space1),
+        (position, parse_identifier),
+    )
+        .parse(s)?;
+    let (_, base_module_name, _, (span, identifier)) = result;
+    let module_name = format!("{}.{}", base_module_name, identifier);
+    Ok((rest, vec![Import::new(
+        module_name,
+        span.location_line(),
+        false,
+    )]))
 }
 
-fn parse_simple_import_statement(pair: Pair<Rule>, context: &mut ParseContext) -> Vec<Import> {
-    let (line_number, _) = pair.line_col();
-
-    pair.into_inner()
-        .flat_map(|inner_pair| match inner_pair.as_rule() {
-            Rule::MODULE => {
-                let imported_object = inner_pair.as_str().to_owned();
-                Some(Import {
-                    imported_object,
-                    line_number: line_number as u32,
-                    typechecking_only: context.typechecking_only,
-                })
-            }
-            Rule::AS_IDENTIFIER => None,
-            _ => unreachable!("{:?}", inner_pair.as_rule()),
-        })
-        .collect()
+fn parse_module(s: Span) -> IResult<Span, &str> {
+    let (rest, result) = recognize(separated_list1(tag("."), parse_identifier)).parse(s)?;
+    Ok((rest, result.fragment()))
 }
 
-fn parse_from_import_statement(pair: Pair<Rule>, context: &mut ParseContext) -> Vec<Import> {
-    let (line_number, _) = pair.line_col();
-    let mut inner_pairs = pair.into_inner();
-    let imported_base = {
-        let mut imported_base = inner_pairs.next().unwrap().as_str();
-        if imported_base.ends_with(".") {
-            imported_base = imported_base.strip_suffix(".").unwrap();
-        }
-        imported_base
-    };
-
-    inner_pairs
-        .filter_map(|inner_pair| match inner_pair.as_rule() {
-            Rule::IDENTIFIER => {
-                let imported_object = format!("{}.{}", imported_base, inner_pair.as_str());
-                Some(Import {
-                    imported_object,
-                    line_number: line_number as u32,
-                    typechecking_only: context.typechecking_only,
-                })
-            }
-            Rule::AS_IDENTIFIER => None,
-            _ => unreachable!("{:?}", inner_pair.as_rule()),
-        })
-        .collect()
+fn parse_identifier(s: Span) -> IResult<Span, &str> {
+    let (rest, result) = recognize(many1(alt((alphanumeric1, tag("_"))))).parse(s)?;
+    Ok((rest, result.fragment()))
 }
 
-fn parse_wildcard_from_import_statement(
-    pair: Pair<Rule>,
-    context: &mut ParseContext,
-) -> Vec<Import> {
-    let (line_number, _) = pair.line_col();
-    let mut inner_pairs = pair.into_inner();
-    let mut imported_l = inner_pairs.next().unwrap().as_str();
-    if imported_l.ends_with(".") {
-        imported_l = imported_l.strip_suffix(".").unwrap();
-    }
-    let imported = format!("{}.*", imported_l);
-    vec![Import {
-        imported_object: imported.to_string(),
-        line_number: line_number as u32,
-        typechecking_only: context.typechecking_only,
-    }]
+fn parse_comment(s: Span) -> IResult<Span, ()> {
+    let (rest, _) = (tag("#"), not_line_ending).parse(s)?;
+    Ok((rest, ()))
 }
 
 #[cfg(test)]
@@ -192,7 +161,7 @@ mod tests {
         ("import  foo  as  FOO ,  bar  as  BAR", &["foo", "bar"]),
         ("import foo # Comment", &["foo"]),
     })]
-    fn test_parse_simple_import_statement(case: (&str, &[&str])) {
+    fn test_parse_import_statement(case: (&str, &[&str])) {
         parse_and_check(case);
     }
