--- src/nom.rs
@@ -1,447 +1,0 @@
-use nom::branch::alt;
-use nom::bytes::complete::tag;
-use nom::character::complete::{alphanumeric1, newline, not_line_ending, space0, space1};
-use nom::combinator::{all_consuming, opt, recognize};
-use nom::multi::{many0, many1, separated_list1};
-use nom::sequence::{delimited, preceded, terminated};
-use nom::{IResult, Parser};
-use nom_locate::{LocatedSpan, position};
-use std::borrow::Borrow;
-
-type Span<'a> = LocatedSpan<&'a str>;
-
-#[derive(Debug, Eq, PartialEq, Clone)]
-pub struct Import {
-    pub imported_object: String,
-    pub line_number: u32,
-    pub typechecking_only: bool,
-}
-
-impl Import {
-    pub fn new<B: Borrow<str>>(
-        imported_object: B,
-        line_number: u32,
-        typechecking_only: bool,
-    ) -> Self {
-        Self {
-            imported_object: imported_object.borrow().to_owned(),
-            line_number,
-            typechecking_only,
-        }
-    }
-}
-
-pub fn parse_imports(s: &str) -> Result<Vec<Import>, String> {
-    let s = Span::new(s);
-    let (_, result) = all_consuming(many0(parse_import_statement_list))
-        .parse(s)
-        .map_err(|e| e.to_string())?;
-    Ok(result.into_iter().flatten().collect())
-}
-
-fn parse_import_statement_list(s: Span) -> IResult<Span, Vec<Import>> {
-    let (rest, result) = terminated(
-        separated_list1(
-            delimited(space0, tag(";"), space0),
-            alt((parse_import_statement, parse_from_import_statement)),
-        ),
-        (
-            opt(space0),
-            opt(tag(";")),
-            opt(space0),
-            opt(parse_comment),
-            opt(newline),
-        ),
-    )
-    .parse(s)?;
-    Ok((rest, result.into_iter().flatten().collect()))
-}
-
-fn parse_import_statement(s: Span) -> IResult<Span, Vec<Import>> {
-    let (rest, result) = preceded(
-        (tag("import"), space1),
-        separated_list1(
-            delimited(space0, tag(","), space0),
-            terminated(
-                (position, parse_module),
-                opt((space1, tag("as"), space1, parse_identifier)),
-            ),
-        ),
-    )
-    .parse(s)?;
-    Ok((
-        rest,
-        result
-            .into_iter()
-            .map(|(span, module_name)| Import::new(module_name, span.location_line(), false))
-            .collect(),
-    ))
-}
-
-fn parse_from_import_statement(s: Span) -> IResult<Span, Vec<Import>> {
-    let (rest, result) = (
-        (tag("from"), space1),
-        parse_module,
-        (space1, tag("import"), space1),
-        (position, parse_identifier),
-    )
-        .parse(s)?;
-    let (_, base_module_name, _, (span, identifier)) = result;
-    let module_name = format!("{}.{}", base_module_name, identifier);
-    Ok((rest, vec![Import::new(
-        module_name,
-        span.location_line(),
-        false,
-    )]))
-}
-
-fn parse_module(s: Span) -> IResult<Span, &str> {
-    let (rest, result) = recognize(separated_list1(tag("."), parse_identifier)).parse(s)?;
-    Ok((rest, result.fragment()))
-}
-
-fn parse_identifier(s: Span) -> IResult<Span, &str> {
-    let (rest, result) = recognize(many1(alt((alphanumeric1, tag("_"))))).parse(s)?;
-    Ok((rest, result.fragment()))
-}
-
-fn parse_comment(s: Span) -> IResult<Span, ()> {
-    let (rest, _) = (tag("#"), not_line_ending).parse(s)?;
-    Ok((rest, ()))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::parse_imports;
-    use parameterized::parameterized;
-
-    #[test]
-    fn test_parse_empty_string() {
-        let imports = parse_imports("").unwrap();
-        assert!(imports.is_empty());
-    }
-
-    fn parse_and_check(case: (&str, &[&str])) {
-        let (code, expected_imports) = case;
-        let imports = parse_imports(code).unwrap();
-        assert_eq!(
-            expected_imports,
-            imports
-                .into_iter()
-                .map(|i| i.imported_object)
-                .collect::<Vec<_>>()
-        );
-    }
-
-    fn parse_and_check_with_typechecking_only(case: (&str, &[(&str, bool)])) {
-        let (code, expected_imports) = case;
-        let imports = parse_imports(code).unwrap();
-        assert_eq!(
-            expected_imports
-                .iter()
-                .map(|i| (i.0.to_owned(), i.1))
-                .collect::<Vec<_>>(),
-            imports
-                .into_iter()
-                .map(|i| (i.imported_object, i.typechecking_only))
-                .collect::<Vec<_>>()
-        );
-    }
-
-    #[parameterized(case = {
-        ("import foo", &["foo"]),
-        ("import foo_FOO_123", &["foo_FOO_123"]),
-        ("import foo.bar", &["foo.bar"]),
-        ("import foo.bar.baz", &["foo.bar.baz"]),
-        ("import foo, bar, bax", &["foo", "bar", "bax"]),
-        ("import foo as FOO", &["foo"]),
-        ("import foo as FOO, bar as BAR", &["foo", "bar"]),
-        ("import  foo  as  FOO ,  bar  as  BAR", &["foo", "bar"]),
-        ("import foo # Comment", &["foo"]),
-    })]
-    fn test_parse_import_statement(case: (&str, &[&str])) {
-        parse_and_check(case);
-    }
-
-    #[parameterized(case = {
-        ("from foo import bar", &["foo.bar"]),
-        ("from foo import bar_BAR_123", &["foo.bar_BAR_123"]),
-        ("from .foo import bar", &[".foo.bar"]),
-        ("from ..foo import bar", &["..foo.bar"]),
-        ("from . import foo", &[".foo"]),
-        ("from .. import foo", &["..foo"]),
-        ("from foo.bar import baz", &["foo.bar.baz"]),
-        ("from .foo.bar import baz", &[".foo.bar.baz"]),
-        ("from ..foo.bar import baz", &["..foo.bar.baz"]),
-        ("from foo import bar, baz, bax", &["foo.bar", "foo.baz", "foo.bax"]),
-        ("from foo import bar as BAR", &["foo.bar"]),
-        ("from foo import bar as BAR, baz as BAZ", &["foo.bar", "foo.baz"]),
-        ("from  foo  import  bar  as  BAR ,  baz  as  BAZ", &["foo.bar", "foo.baz"]),
-        ("from foo import bar # Comment", &["foo.bar"]),
-    })]
-    fn test_parse_from_import_statement(case: (&str, &[&str])) {
-        parse_and_check(case);
-    }
-
-    #[parameterized(case = {
-        ("from foo import (bar)", &["foo.bar"]),
-        ("from foo import (bar,)", &["foo.bar"]),
-        ("from foo import (bar, baz)", &["foo.bar", "foo.baz"]),
-        ("from foo import (bar, baz,)", &["foo.bar", "foo.baz"]),
-        ("from foo import (bar as BAR, baz as BAZ,)", &["foo.bar", "foo.baz"]),
-        ("from  foo  import  ( bar  as  BAR , baz  as  BAZ , )", &["foo.bar", "foo.baz"]),
-        ("from foo import (bar, baz,) # Comment", &["foo.bar", "foo.baz"]),
-
-        (r#"
-from foo import (
-    bar,
-    baz
-)
-        "#, &["foo.bar", "foo.baz"]),
-
-        (r#"
-from foo import (
-    bar,
-    baz,
-)
-        "#, &["foo.bar", "foo.baz"]),
-
-        (r#"
-from foo import (
-    a, b,
-    c, d,
-)
-        "#, &["foo.a", "foo.b", "foo.c", "foo.d"]),
-
-        // As name
-        (r#"
-from foo import (
-    bar as BAR,
-    baz as BAZ,
-)
-        "#, &["foo.bar", "foo.baz"]),
-
-        // Whitespace
-        (r#"
-from  foo  import  (
-
-    bar  as  BAR ,
-
-       baz  as  BAZ ,
-
-)
-        "#, &["foo.bar", "foo.baz"]),
-
-        // Comments
-        (r#"
-from foo import ( # C
-    # C
-    bar as BAR, # C
-    # C
-    baz as BAZ, # C
-    # C
-) # C
-        "#, &["foo.bar", "foo.baz"]),
-    })]
-    fn test_parse_multiline_from_import_statement(case: (&str, &[&str])) {
-        parse_and_check(case);
-    }
-
-    #[parameterized(case = {
-        ("from foo import *", &["foo.*"]),
-        ("from .foo import *", &[".foo.*"]),
-        ("from ..foo import *", &["..foo.*"]),
-        ("from . import *", &[".*"]),
-        ("from .. import *", &["..*"]),
-        ("from  foo  import  *", &["foo.*"]),
-        ("from foo import * # Comment", &["foo.*"]),
-    })]
-    fn test_parse_wildcard_from_import_statement(case: (&str, &[&str])) {
-        parse_and_check(case);
-    }
-
-    #[parameterized(case = {
-        ("import a; import b", &["a", "b"]),
-        ("import a; import b;", &["a", "b"]),
-        ("import  a ;  import  b ;", &["a", "b"]),
-        ("import a; import b # Comment", &["a", "b"]),
-        ("import a; from b import c; from d import (e); from f import *", &["a", "b.c", "d.e", "f.*"]),
-    })]
-    fn test_parse_import_statement_list(case: (&str, &[&str])) {
-        parse_and_check(case);
-    }
-
-    #[parameterized(case = {
-        (r#"
-import a, b, \
-       c, d
-        "#, &["a", "b", "c", "d"]),
-
-        (r#"
-from foo import a, b, \
-                c, d
-        "#, &["foo.a", "foo.b", "foo.c", "foo.d"]),
-
-        (r#"
-from foo \
-    import *
-        "#, &["foo.*"]),
-    })]
-    fn test_backslash_continuation(case: (&str, &[&str])) {
-        parse_and_check(case);
-    }
-
-    #[parameterized(case = {
-        (r#"
-import a
-def foo():
-    import b 
-import c
-        "#, &["a", "b", "c"]),
-
-        (r#"
-import a
-class Foo:
-    import b
-import c
-        "#, &["a", "b", "c"]),
-    })]
-    fn test_parse_nested_imports(case: (&str, &[&str])) {
-        parse_and_check(case);
-    }
-
-    #[parameterized(case = {
-        (r#"
-import foo
-if typing.TYPE_CHECKING: import bar
-import baz
-"#, &[("foo", false), ("bar", true), ("baz", false)]),
-
-        (r#"
-import foo
-if TYPE_CHECKING: import bar
-import baz
-"#, &[("foo", false), ("bar", true), ("baz", false)]),
-
-        (r#"
-import foo
-if  TYPE_CHECKING :  import bar
-import baz
-"#, &[("foo", false), ("bar", true), ("baz", false)]),
-
-        (r#"
-import foo
-if TYPE_CHECKING: import bar as BAR
-import baz
-"#, &[("foo", false), ("bar", true), ("baz", false)]),
-
-        (r#"
-import foo # C
-if TYPE_CHECKING: import bar # C
-import baz # C
-"#, &[("foo", false), ("bar", true), ("baz", false)]),
-    })]
-    fn test_singleline_if_typechecking(case: (&str, &[(&str, bool)])) {
-        parse_and_check_with_typechecking_only(case);
-    }
-
-    #[parameterized(case = {
-        (r#"
-import foo
-if typing.TYPE_CHECKING:
-    import bar
-import baz
-"#, &[("foo", false), ("bar", true), ("baz", false)]),
-
-        (r#"
-import foo
-if TYPE_CHECKING:
-    import bar
-import baz
-"#, &[("foo", false), ("bar", true), ("baz", false)]),
-
-        (r#"
-import  foo
-
-if  TYPE_CHECKING :
-
-    import  bar
-
-import  baz
-"#, &[("foo", false), ("bar", true), ("baz", false)]),
-
-        (r#"
-import foo
-if TYPE_CHECKING:
-    import bar as BAR
-import baz
-"#, &[("foo", false), ("bar", true), ("baz", false)]),
-
-        (r#"
-import foo # C
-if TYPE_CHECKING: # C
-    # C
-    import bar # C
-    # C
-import baz # C
-"#, &[("foo", false), ("bar", true), ("baz", false)]),
-
-        (r#"
-import foo
-if TYPE_CHECKING:
-    """
-    Comment
-    """
-    import bar
-import baz
-"#, &[("foo", false), ("bar", true), ("baz", false)]),
-    })]
-    fn test_multiline_if_typechecking(case: (&str, &[(&str, bool)])) {
-        parse_and_check_with_typechecking_only(case);
-    }
-
-    #[parameterized(case = {
-        (r#"
-import foo
-"""
-import bar
-"""
-import baz
-"#, &["foo", "baz"]),
-
-        (r#"
-import foo
-'''
-import bar
-'''
-import baz
-"#, &["foo", "baz"]),
-    })]
-    fn test_multiline_strings(case: (&str, &[&str])) {
-        parse_and_check(case);
-    }
-
-    #[test]
-    fn test_parse_line_numbers() {
-        let imports = parse_imports(
-            "
-import a
-from b import c
-from d import (e)
-from f import *",
-        )
-        .unwrap();
-        assert_eq!(
-            vec![
-                ("a".to_owned(), 2_u32),
-                ("b.c".to_owned(), 3_u32),
-                ("d.e".to_owned(), 4_u32),
-                ("f.*".to_owned(), 5_u32),
-            ],
-            imports
-                .into_iter()
-                .map(|i| (i.imported_object, i.line_number))
-                .collect::<Vec<_>>()
-        );
-    }
-}
